Expose and test makeRecencyFilter so cgm-remote-monitor can use it
