+++ index.js
@@ -1,5 +1,6 @@
 module.exports = {
   carelink: require('./carelink'),
+  filter: require('./filter'),
   logger: require('./logger'),
   nightscout: require('./nightscout'),
   transform: require('./transform')
